requirements.txt
PyVISA==1.11.3
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.2
visa==2.21.0
